refactorointia ja testejä backendille
