+++ package.json
@@ -18,14 +18,14 @@
     "Ville Salmela",
     "Ville Suonurmi"
   ],
-  "type": "commonjs",
+  "type": "module",
   "main": "index.ts",
   "scripts": {
-    "start": "ts-node index.ts",
-    "dev": "ts-node-dev index.ts",
+    "start": "tsx index.ts",
+    "dev": "tsx --watch index.ts",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "tsx test/*.test.ts"
   },
   "dependencies": {
     "axios": "^1.9.0",
@@ -34,15 +34,17 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.26.0",
+    "@types/axios": "^0.9.36",
     "@types/express": "^5.0.1",
     "@types/node": "^22.15.18",
+    "@types/supertest": "^6.0.3",
     "@types/xml2js": "^0.4.14",
     "@typescript-eslint/eslint-plugin": "^8.32.1",
     "@typescript-eslint/parser": "^8.32.1",
     "eslint": "^9.26.0",
     "globals": "^16.1.0",
-    "ts-node": "^10.9.2",
-    "ts-node-dev": "^2.0.0",
+    "supertest": "^7.1.1",
+    "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.32.1"
   }
